# api.yml
openapi: 3.0.3
info:
  title: trainingApp API
  description: Сервис учёта тренировок и аналитики нагрузки.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Локальная разработка
...
paths:
...
  /api/v1/exercises:
    get:
      tags: [exercises]
      summary: Каталог упражнений
      description: Возвращает системные упражнения и личные упражнения текущего пользователя.
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: Список упражнений
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Exercise" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/Internal" }
    post:
      tags: [exercises]
      summary: Создать личное упражнение
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExerciseRequest" }
      responses:
        "201":
          description: Упражнение создано
          headers:
            Location:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Exercise" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "422": { $ref: "#/components/responses/Unprocessable" }
        "500": { $ref: "#/components/responses/Internal" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Токен отсутствует, некорректен или истёк
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Unprocessable:
      description: Запрос корректен синтаксически, но нарушает правила предметной области
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Internal:
      description: Внутренняя ошибка сервиса
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    ExerciseRequest:
      type: object
      properties:
        name: { type: string, maxLength: 100 }
        description: { type: string, maxLength: 1000 }
      required: [name]
    Exercise:
      type: object
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        description: { type: string }
        user_id:
          type: integer
          format: int64
          nullable: true
          description: null — системное упражнение, доступное всем пользователям.
        created_at: { type: string, format: date-time }
...
security:
  - bearerAuth: []